requirements.txt
PyQt5==5.15.11
PyQt5-sip==12.15.0
matplotlib==3.9.2
supabase==2.5.3
postgrest==0.13.0
gotrue==2.5.0
python-dotenv==1.0.1
cryptography==43.0.1
requests==2.32.3
httpx==0.27.2
realtime==1.0.6
websockets==12.0